+++ backend/internal/service/handler/summary/get_gross_summary.go
@@ -32,31 +32,3 @@ func (h *Handler) GetGrossSummary(c *fiber.Ctx) error {
 
 	return c.Status(fiber.StatusOK).JSON(grossSummary)
 }
-
-func (h *Handler) GetNetSummary(c *fiber.Ctx) error {
-	companyID := c.Query("company_id")
-	startDate := c.Query("start_date")
-	endDate := c.Query("end_date")
-
-	if companyID == "" {
-		return errs.BadRequest("Company ID is required")
-	}
-	if startDate == "" {
-		return errs.BadRequest("Start Date is required")
-	}
-	if endDate == "" {
-		return errs.BadRequest("End Date is required")
-	}
-
-	netSummary, err := h.summaryRepository.GetNetSummary(
-		c.Context(),
-		companyID,
-		startDate,
-		endDate,
-	)
-	if err != nil {
-		return err
-	}
-
-	return c.Status(fiber.StatusOK).JSON(netSummary)
-}
